sysfs-i2c and bcm283x: Add support to change I²C bus speed on the fly
This may enable an app (running as root) to boost the speed to 400kHz, which is
sometimes a requirement.

Sadly reading the current value also requires root so this cannot be skipped if
the current value is the desired one. Also on bcm283x, the default value isn't
set correctly by the driver.