cmd: enable optional use of periph.io/x/extra/hostextra.Init()
This is done via:
  go install -tags periphextra periph.io/x/periph/cmd/...
or
  go get -u -tags periphextra periph.io/x/periph/cmd/...

If the tag `periphextra` is not provided, periph continues to not depend on
third party library.

Manually confirmed that "go get -u periph.io/x/periph/cmd/..." does not download
undesired dependencies with this change.

This enables having only one source file for the executables in cmd, while
enabling optional compilation of extended drivers for users with a simple tag.

The exception will be periph-smoketest, which will be copied pasted into
periph.io/x/extra/cmd/extra-smoketest in due time (when the first smoketest is
written there).

The rationale is that the build 'periphextra' only add dependencies on
hostextra.Init(), nothing else. In particular smoke tests can come and go, and
keeping the list of smoketest located in extra in periph-smoketest up to date
would cause a fair amount of pain.

Instead of making a wrapper function, the other investigated idea has been to
do:
  var hostInit = host.Init (or hostextra.Init)

The main downside of this is that the reference becomes a non-constant, and I'm
not sure the compiler optimizes this.