periph-web: use reflection to make the functions more intuitive
This feels more like a dynamic language! At a relatively small performance cost.

More enforcements:

On Go 1.10+, unknown fields in the request are detected and refused as HTTP 400.
On Go 1.9-, unknown fields in the request are silently ignored.

Enforce that no content is sent on GET.

Enforce Content-Type application/json on POST.